pyproject.toml
[tool.poetry]
name = "object-diff"
version = "0.1.0"
description = ""
authors = ["Ryusei Kakujo <ryusei.kakujo@traffine.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.4"
requests = "^2.31.0"
joblib = "^1.3.2"
mangum = "^0.17.0"
pillow = "^10.2.0"
python-multipart = "^0.0.9"
google-cloud-vision = "^3.7.2"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
torch = "^2.2.2"
torchvision = "^0.17.2"
uvicorn = "^0.29.0"
fastapi = "^0.110.0"
httpx = "^0.27.0"
loguru = "^0.7.2"
onnxruntime = "<=1.15.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ipdb = "^0.13.13"
pysen = {version = "^0.11.0", extras = ["lint"]}
tf2onnx = "^1.16.1"
tensorflow = "^2.16.1"
streamlit = "^1.32.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black] # automatically generated by pysen
line-length = 250
target-version = ["py310"]

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 250
multi_line_output = 3
use_parentheses = true

[tool.pysen]
version = "0.11.0"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py310"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]
